Move bs2tick and co. to a new file
This non-functional change moves these bs2-prefixed functions to a new file (without changing their contents).

As well as splitting this code out from `main.c`, it allows common constants and the device state to be made accessible from its header file.
